requirements.txt
google-adk
feedparser